Fixes #122
There is still an open issue around Case B1 for which I'll put in a new ticket.

The commit creates a national default language which is returned if no circle is provided.  It also returns all the valid languages for a circle if the user has no language preference saved.