Fixes NIP-34
This allows for the creation and association of States, Districts, Circles and Langauges.  I was able to create a subscriber, but the number of related tables made creation of all fields through the UI problematic.  There isn't a requirement for creation of this data through the UI though so I don't feel like we need to fix all of those issues.