Fixes NMS-236
When a user is marked as INVALID we now null out their contact number.  The non-null
constraint in the database has been removed but a class level validator has been added that
enforces a contact number when the status is not INVALID and that contact number is null when
status is INVALID.

Since the INVALID FLW has no contact number there will be no conflict if a user with that number calls MA/MK.

All other information is retained in the record.